redundency checks

+++ source/functions/append.py
@@ -1,19 +1,9 @@
-
-# 	answer Function temp readme
-#	
-#	appends a new answer result to a question for a specified survey
-# 	
-#	Note that this expects a survey file and will error if missing
-#
-# 	Example CSV File Format:
-#
-#	question_id,answers
-#	1,"['a', 'a', 'b', 'b', 'c', 'a', 'd', 'j']"
-#
-# 	append.answer("test_answers","1","j") -- will open file test_answers.csv, 
-# 	and append the answer j to question 1
-
 import csv, ast, os
+from classes import fileclasses
+from functions import append
+from server import errorMSG
+
+
 
 #####################################################################
 # 					append.answer()									#
@@ -28,23 +18,38 @@ import csv, ast, os
 # Sample CSV Format: 1,"['a', 'a', 'b', 'b', 'c', 'a', 'd', 'j']" 	#
 #####################################################################
 
-def answer(s_id, q_id, answer):
-	file = str(s_id)+".csv"
-	tmp = str(s_id)+"_tmp.csv"	
-	with open(file, 'r+') as csvReadFile:
-		fieldnames = ['question_id', 'answers']
-		reader = csv.DictReader(csvReadFile, fieldnames=fieldnames)
-		with open (tmp, 'w') as write_row:
-			writer=csv.DictWriter(write_row, fieldnames=fieldnames)
-			for row in reader:
-				if(row['question_id']==q_id):
-					tmp_list=ast.literal_eval(row['answers'])
-					tmp_list.append(answer)
-					row['answers'] = tmp_list
-				writer.writerow(row)				
-	os.remove(file)
-	os.rename(tmp, file)
+def answer(sID, qID, answer):
 
+	if(sID==""):
+		errorMSG("append.answer","No sID Provided")
+		return
+	if(qID==""):
+		errorMSG("append.answer","No qID Provided")
+		return
+	if(answer==""):
+		errorMSG("append.answer","No Answer Provided")
+		return
+
+	file = str(sID)+".csv"
+	tmp = str(sID)+"_tmp.csv"	
+
+	try:
+		with open(file, 'r+') as csvReadFile:
+			fieldnames = ['question_id', 'answers']
+			reader = csv.DictReader(csvReadFile, fieldnames=fieldnames)
+			with open (tmp, 'w') as write_row:
+				writer=csv.DictWriter(write_row, fieldnames=fieldnames)
+				for row in reader:
+					if(row['question_id']==qID):
+						tmp_list=ast.literal_eval(row['answers'])
+						tmp_list.append(answer)
+						row['answers'] = tmp_list
+					writer.writerow(row)				
+		os.remove(file)
+		os.rename(tmp, file)
+	except IOError:
+		errorMSG("append.answer","No Survey Results File Found")
+		return	
 
 #####################################################
 # 				append.question() 					#
@@ -52,10 +57,23 @@ def answer(s_id, q_id, answer):
 
 def question(sID, question, answers): #assumes all params are strings
 
-	master_question(qID, question, my_answers)
+	if(sID==""):
+		errorMSG("append.question","No sID Provided")
+		return
+	if(question==""):
+		errorMSG("append.question","No Question Provided")
+		return
+	if(answers==""):
+		errorMSG("append.question","No Answers Provided")
+		return
+
+	ID = fileclasses.textfile("questionID.txt")
+	qID = ID.updateID()
+
+	append.master_question(qID, question, answers)
 
 	if (sID != -1): # -1 is a flag meaning there is no survey to immediately add it to
-		master_survey(qID, sID)
+		append.master_survey(qID, sID)
 
 
 #####################################################
@@ -64,16 +82,30 @@ def question(sID, question, answers): #assumes all params are strings
 
 def master_question(qID, question, answers):
 
+	if(qID==""):
+		errorMSG("append.master_question","No qID Provided")
+		return
+	if (question==""):
+		errorMSG("append.master_question","No Question Provided")
+		return
+	if(answers==""):
+		errorMSG("append.master_question","No Answers Provided")
+		return
+
 	# write new row in order of qID, question, answers (in form id, question, ans1, ans2, ans3 etc.)
 	with open('question_temp.csv','w+', newline = '') as csv_out:
 			writer = csv.writer(csv_out)
 			writer.writerow([qID, question, answers])
 
 	# overwrite master with changes and get rid of "" symbols
-	with open('question_temp.csv') as csv_in, open('master_question.csv', 'a') as csv_out:
-		for line in csv_in:
-			csv_out.write(line.replace('\"',''))
-
+	try:
+		with open('question_temp.csv') as csv_in, open('master_question.csv', 'a') as csv_out:
+			for line in csv_in:
+				csv_out.write(line.replace('\"',''))
+		os.remove('question_temp.csv')
+	except IOError:
+		errorMSG("append.master_question","Error overwriting master with changes")
+		return	
 
 #####################################################
 # 				append.master_survey()				#
@@ -81,8 +113,19 @@ def master_question(qID, question, answers):
 
 def master_survey(qID, sID):
 
+	if(qID==""):
+		errorMSG("append.master_survey","No qID Provided")
+		return
+	if (sID==""):
+		errorMSG("append.master_survey","No sID Provided")
+		return
+
 	# get current info in master_survey.csv file
-	reader = csv.reader(open('master_survey.csv'))
+	try:
+		reader = csv.reader(open('master_survey.csv'))
+	except IOError:
+		errorMSG("append.master_survey","File master_survey.csv not found")
+		return
 	survey_list = []
 	for row in reader:
 		current_survey = {
@@ -118,6 +161,11 @@ def master_survey(qID, sID):
 		dict_writer.writerows(survey_list)
 
 	# write from temp back to master_survey file to write in changes (without quotations)
-	with open('survey_temp.csv') as csv_in, open('master_survey.csv', 'w') as csv_out:
-		for line in csv_in:
-			csv_out.write(line.replace('\"',''))
+	try:
+		with open('survey_temp.csv') as csv_in, open('master_survey.csv', 'w') as csv_out:
+			for line in csv_in:
+				csv_out.write(line.replace('\"',''))
+		os.remove('survey_temp.csv')
+	except IOError:
+		errorMSG("append.master_survey","Error overwriting master_survey file with changes")
+		return
